Add MSTS and MSTR microcode instructions
This establishes an internal status register (MST) that is affected by micro codes which can then be returned in whole or in part to the public accessible ST register. This is done through two new microcodes:
- The MSTS microcode (Microcode STatus Set) allows direct clearing and setting of any subset of ZSCO status flags for the internal MST register.
- The MSTR microcode (Microcode STatus Return) support propagating clears or sets in the MST register to the public ST register.

This also adds full documentation for the existing microcode operations and documents the syntax of microcode assembly.