Overhaul registers
Removed DP register, and added "base" memory registers for each bank: BC, BS, BD, and BE.

Got rid of the SD register, it was a weird anomaly.

Added C2 back in because having an odd number of registers is unappealing.

Moved C0, C1, C2, and ST to the end of the list. This puts all of them outside the normal  addresable range with the ArgType::kWordReg argument. They are still addressible in microcode, so that means they could be exposed -- but it would be pointless as you can't write to them with any persistence.